Move learningpath meta image logic to coverphoto resolver
