+++ src/api/folderResourceMetaApi.ts
@@ -10,6 +10,7 @@ import groupBy from "lodash/groupBy";
 import { IArticleV2DTO } from "@ndla/types-backend/article-api";
 import { IAudioMetaInformationDTO } from "@ndla/types-backend/audio-api";
 import { IImageMetaInformationV2DTO } from "@ndla/types-backend/image-api";
+import { ILearningPathV2DTO } from "@ndla/types-backend/learningpath-api";
 import { ResourceType } from "@ndla/types-backend/myndla-api";
 import { Node } from "@ndla/types-taxonomy";
 import { fetchAudio } from "./audioApi";
@@ -26,7 +27,6 @@ import {
   GQLQueryFolderResourceMetaSearchArgs,
 } from "../types/schema";
 import { articleToMeta, learningpathToMeta } from "../utils/apiHelpers";
-import { isNDLAEmbedUrl } from "../utils/articleHelpers";
 
 const findResourceTypes = (result: Node | null, context: ContextWithLoaders): GQLFolderResourceResourceType[] => {
   const ctx = result?.contexts?.[0];
@@ -45,24 +45,9 @@ const fetchResourceMeta = async (
 ): Promise<Array<GQLMeta | undefined>> => {
   if (type === "learningpath") {
     const learningpaths = await context.loaders.learningpathsLoader.loadMany(ids);
-    return Promise.all(
-      learningpaths
-        .filter((learningpath) => !!learningpath)
-        .map(async (learningpath) => {
-          const learningStepWithResource = learningpath.learningsteps.find(
-            (learningStep) =>
-              learningStep.embedUrl &&
-              learningStep.embedUrl.url &&
-              (learningStep.embedUrl.embedType === "oembed" || learningStep.embedUrl.embedType === "iframe") &&
-              isNDLAEmbedUrl(learningStep.embedUrl.url),
-          );
-
-          const articleId = learningStepWithResource?.embedUrl?.url.split("/").pop();
-          const article = articleId ? await context.loaders.articlesLoader.load(articleId) : undefined;
-
-          return learningpathToMeta(learningpath, article?.metaImage);
-        }),
-    );
+    return learningpaths
+      .filter((learningpath): learningpath is ILearningPathV2DTO => !!learningpath)
+      .map(learningpathToMeta);
   } else {
     const articles = await context.loaders.articlesLoader.loadMany(ids);
     return articles.filter((article): article is IArticleV2DTO => !!article).map(articleToMeta);
